increase vec2 and mat3 precision with float64

+++ src/math/mat3.ts
@@ -7,7 +7,7 @@ const IDENTITY = [
   0, 0, 1
 ];
 
-export class Matrix3 extends Float32Array {
+export class Matrix3 extends Float64Array {
   constructor() {
     super(9);
     this.set(IDENTITY);
